vheaders: rewrite MakeSDN to implement "global" flow table
This version of MakeSDN uses an `impl` type that is a flowtable which
should be applied to all switches and a map from switches to their
specific flowtables. Switch-specific flowtables already incorporate the
global policy.

This version also uses the local policy created by the compiler to
collect the virtual headers and associated values according to whether
or not they have a realized header. This result is then used by the
devirtualize function to build a total mapping from virtual headers to
realizable headers, and virtual headers and values to values. The code
passes these two maps around to convert Patters and Actions
accoridingly.